finish tooltip
