+++ frontend/src/page/MainLayout.tsx
@@ -1,11 +1,11 @@
 import { Box, Toolbar } from '@mui/material';
 import { NavigationBar } from '../components/NavigationBar';
 import { useEffect, useState } from 'react';
-import Overview from './Overview';
 import GatekeeperLayout from './GatekeeperLayout';
 import { useNavigate, useOutlet, useParams } from 'react-router-dom';
 import { useMsal } from '@azure/msal-react';
 import { routes } from '../utils/route';
+import Overview from './Overview';
 
 const MainLayout = () => {
   const [page, setPage] = useState(0);
