+++ frontend/src/styles/index.ts
@@ -177,6 +177,21 @@ export const getDesignTokens = (mode: PaletteMode): ThemeOptions => ({
         },
       },
     },
+    MuiTooltip: {
+      styleOverrides: {
+        tooltip: {
+          backgroundColor: '#fff',
+          color: '#000',
+          border: '1px solid var(--sunglow-600, #FBBB21)',
+          borderRadius: 0,
+          padding: '6px 8px',
+          fontSize: '10px',
+          fontWeight: 400,
+          lineHeight: '16px',
+          maxWidth: 'none',
+        },
+      },
+    },
     // ...(mode === 'light' ? {} : {}),
   },
   typography: {
@@ -194,5 +209,20 @@ export const globalStyles = {
     fontSize: '32px',
     fontWeight: 500,
     color: 'inherit',
+  },
+  drawerInputLabel: {
+    fontSize: '14px',
+    fontWeight: 400,
+    width: '64px',
+  },
+  drawerLabelContent: {
+    fontSize: '14px',
+    fontWeight: 400,
+  },
+  buttonTinySize: {
+    padding: '6px 8px',
+    fontWeight: 400,
+    fontSize: '12px',
+    lineHeight: '16px',
   },
 };
